test-requirements.txt
typing
typing-extensions
mock
coveralls
django-storages
boto3
boto
google-cloud-storage
pytest
pytest-django
